requirements (2024_07_05 16_18_15 UTC).txt
# requirements.txt
torch
torchvision
transformers
flask
gradio
requests
Pillow
